requirements.txt
jieba==0.39
stanfordcorenlp
jpype1